requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
bleak==3.0.1
pytz==2025.2
python-dotenv==1.1.1
